created separate graph testing module
